Issue#69 detailed flight class

See merge request comp3350-winter2026/a02-g09-theincredibles!40

+++ app/src/main/java/com/group9/ongo/business/services/BookingServiceImpl.java
@@ -9,6 +9,7 @@ import com.group9.ongo.models.BookingDetails;
 import com.group9.ongo.models.Flight;
 import com.group9.ongo.models.Passenger;
 import com.group9.ongo.models.PassengerInput;
+import com.group9.ongo.models.Seat;
 import com.group9.ongo.persistence.BookingRepository;
 import com.group9.ongo.persistence.PassengerRepository;
 
@@ -21,33 +22,49 @@ public class BookingServiceImpl implements BookingService {
     private PassengerRepository passengerRepo;
 
     private FlightService flightService;
+    private SeatService seatService;
+
 
     public BookingServiceImpl(BookingRepository bookingsRepo, PassengerRepository passengerRepo,
-                              FlightService flightService) {
+                              FlightService flightService, SeatService seatService) {
         this.bookingsRepo = bookingsRepo;
         this.passengerRepo = passengerRepo;
         this.flightService = flightService;
+        this.seatService = seatService;
     }
 
     public List<Booking> getBookingByUserId(int userId) {
         return bookingsRepo.getBookingByUserId(userId);
     }
 
-    public Booking createBooking(int userId, int flightId, PassengerInput passengerInfo) throws BookingException, ValidationException {
+    public Booking createBooking(int userId, int flightId, PassengerInput passengerInfo, int seatRow, String seatColumn) throws BookingException, ValidationException {
         BookingValidator.validate(flightService.getFlightById(flightId), passengerInfo);
-        Booking booking = bookingsRepo.addBooking(new Booking(0, userId, flightId));
+
+        Seat seat = seatService.findSeat(flightId, seatRow, seatColumn);
+        Booking booking = bookingsRepo.addBooking(new Booking(0, userId, flightId, seat.getSeatId()));
         Passenger passenger = passengerRepo.addPassenger(passengerInfo, booking.getBookingId());
+        seatService.bookSeat(flightId, seat.getSeatId());
 
         if (passenger == null) {
             bookingsRepo.deleteBooking(booking.getBookingId());
+            seatService.unbookSeat(flightId, seat.getSeatId());
             throw new RuntimeException(BOOKING_PASSENGER_ERROR);
         }
 
         return booking;
     }
 
-    public boolean cancelBooking(int bookingId) {
+    public boolean cancelBooking(int bookingId) throws ValidationException {
+        Booking booking = bookingsRepo.getBookingById(bookingId);
+
+        if(booking == null) return false;
+
         passengerRepo.deletePassengersByBookingId(bookingId);
+
+        Seat seat = seatService.getSeatById(booking.getFlightId(), booking.getSeatId());
+
+        seatService.unbookSeat(seat.getFlightId(), seat.getSeatId());
+
         return bookingsRepo.deleteBooking(bookingId);
     }
